# Conflicts:
#	.travis.yml
#	README.md
#	baystation12.dme
#	code/__defines/_planes+layers.dm
#	code/__defines/atmos.dm
#	code/__defines/guns.dm
#	code/_helpers/lists.dm
#	code/_helpers/text.dm
#	code/_onclick/other_mobs.dm
#	code/datums/extensions/armor/armor.dm
#	code/game/antagonist/antagonist.dm
#	code/game/gamemodes/events/clang.dm
#	code/game/gamemodes/mixed/intrigue.dm
#	code/game/gamemodes/mixed/lizard.dm
#	code/game/gamemodes/mixed/unity.dm
#	code/game/gamemodes/mixed/visitors.dm
#	code/game/machinery/Sleeper.dm
#	code/game/machinery/cryo.dm
#	code/game/machinery/doors/airlock.dm
#	code/game/machinery/doors/firedoor_assembly.dm
#	code/game/machinery/telecomms/telecomunications.dm
#	code/game/machinery/vending.dm
#	code/game/objects/items/devices/scanners/_scanner.dm
#	code/game/objects/items/devices/scanners/gas.dm
#	code/game/objects/items/weapons/candle.dm
#	code/game/objects/items/weapons/grenades/chem_grenade.dm
#	code/game/objects/items/weapons/grenades/light.dm
#	code/game/objects/items/weapons/lighter.dm
#	code/game/objects/objs.dm
#	code/game/objects/structures/crates_lockers/closets/__closet.dm
#	code/game/objects/structures/displaycase.dm
#	code/game/objects/structures/signs.dm
#	code/game/objects/structures/watercloset.dm
#	code/game/objects/structures/window.dm
#	code/game/sound.dm
#	code/modules/admin/admin_verbs.dm
#	code/modules/ascent/ascent_rigs.dm
#	code/modules/bsa/bsa_fire.dm
#	code/modules/client/client_procs.dm
#	code/modules/clothing/head/helmet.dm
#	code/modules/clothing/masks/gasmask.dm
#	code/modules/clothing/spacesuits/rig/suits/station.dm
#	code/modules/clothing/spacesuits/syndi.dm
#	code/modules/clothing/spacesuits/void/station.dm
#	code/modules/clothing/suits/jobs.dm
#	code/modules/clothing/suits/utility.dm
#	code/modules/clothing/suits/wiz_robe.dm
#	code/modules/clothing/under/accessories/armor.dm
#	code/modules/codex/entries/paperwork.dm
#	code/modules/customitems/item_spawning.dm
#	code/modules/emotes/definitions/visible.dm
#	code/modules/events/computer_damage.dm
#	code/modules/food/replicator.dm
#	code/modules/hydroponics/seed.dm
#	code/modules/hydroponics/trays/tray.dm
#	code/modules/integrated_electronics/core/assemblies.dm
#	code/modules/materials/definitions/materials_stone.dm
#	code/modules/mob/grab/grab_datum.dm
#	code/modules/mob/living/carbon/brain/MMI.dm
#	code/modules/mob/living/carbon/human/human.dm
#	code/modules/mob/living/carbon/human/human_defense.dm
#	code/modules/mob/living/carbon/human/unarmed_attack.dm
#	code/modules/mob/living/living.dm
#	code/modules/mob/living/living_defense.dm
#	code/modules/mob/living/silicon/pai/pai.dm
#	code/modules/mob/living/simple_animal/aquatic/_aquatic.dm
#	code/modules/mob/living/simple_animal/hostile/hivebot.dm
#	code/modules/mob/new_player/new_player.dm
#	code/modules/modular_computers/computers/modular_computer/interaction.dm
#	code/modules/modular_computers/computers/modular_computer/power.dm
#	code/modules/modular_computers/computers/subtypes/dev_console.dm
#	code/modules/modular_computers/computers/subtypes/preset_console.dm
#	code/modules/modular_computers/file_system/programs/antagonist/dos.dm
#	code/modules/modular_computers/file_system/programs/antagonist/uplink.dm
#	code/modules/modular_computers/file_system/programs/command/card.dm
#	code/modules/modular_computers/file_system/programs/generic/email_client.dm
#	code/modules/modular_computers/file_system/programs/generic/file_browser.dm
#	code/modules/modular_computers/file_system/programs/generic/wordprocessor.dm
#	code/modules/modular_computers/ntos/ui.dm
#	code/modules/modular_computers/terminal/terminal.dm
#	code/modules/modular_computers/terminal/terminal_commands.dm
#	code/modules/organs/external/_external.dm
#	code/modules/organs/external/_external_damage.dm
#	code/modules/paperwork/adminpaper.dm
#	code/modules/power/lighting.dm
#	code/modules/power/solar.dm
#	code/modules/projectiles/gun.dm
#	code/modules/projectiles/guns/energy/secure.dm
#	code/modules/projectiles/guns/energy/stun.dm
#	code/modules/projectiles/guns/launcher/rocket.dm
#	code/modules/projectiles/guns/magnetic/magnetic_railgun.dm
#	code/modules/projectiles/guns/projectile/automatic.dm
#	code/modules/projectiles/guns/projectile/shotgun.dm
#	code/modules/reagents/Chemistry-Grinder.dm
#	code/modules/reagents/dispenser/dispenser_presets.dm
#	code/modules/reagents/reagent_containers/food/drinks/bottle.dm
#	code/modules/species/outsider/vox.dm
#	code/modules/species/station/station.dm
#	code/modules/sprite_accessories/_accessory_markings.dm
#	code/modules/xenoarcheaology/tools/tools.dm
#	code/procs/dbcore.dm
#	icons/atmos/vent_pump.dmi
#	icons/atmos/vent_scrubber.dmi
#	icons/misc/buildmode.dmi
#	maps/random_ruins/exoplanet_ruins/ec_old_crash/ec_old_crash.dmm
#	maps/torch/items/clothing/solgov-suit.dm
#	maps/torch/job/torch_jobs.dm
#	maps/torch/z1_admin.dmm
#	test/check-paths.sh
#	test/run-test.sh

+++ code/modules/admin/verbs/randomverbs.dm
@@ -625,7 +625,13 @@ Traitors and the like can also be revived with the previous role mostly intact.
 		log_admin("[key_name(usr)] deleted [O] at ([O.x],[O.y],[O.z])")
 		message_admins("[key_name_admin(usr)] deleted [O] at ([O.x],[O.y],[O.z])", 1)
 		SSstatistics.add_field_details("admin_verb","DEL") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
-		qdel(O)
+
+		// turfs are special snowflakes that'll explode if qdel'd
+		if (isturf(O))
+			var/turf/T = O
+			T.ChangeTurf(world.turf)
+		else
+			qdel(O)
 
 /client/proc/cmd_admin_list_open_jobs()
 	set category = "Admin"
